services/dast-scanner/requirements.txt
nats-py==2.7.2
httpx==0.27.0
python-dotenv==1.0.1
prometheus-client==0.20.0
